Implementation : Fee review changes. M2 code.Test cases modified.
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@10190 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/fees/struts/actionforms/FeeActionForm.java
@@ -308,7 +308,7 @@ public class FeeActionForm extends ValidatorActionForm {
 	}
 
 	private boolean isCategoryLoan() {
-		return Short.valueOf(FeeCategory.LOAN.getValue()).equals(
+		return FeeCategory.LOAN.getValue().equals(
 				categoryType.getCategoryId());
 	}
 }
